Rename src/bark.js -> src/index.js
